Move BusinessTypeOptions into business-form-types, remove structure options
