--- packages/webcomponents/src/components/business-forms/utils/business-form-select-options.ts
@@ -1,30 +1,0 @@
-type BusinessType =
-  | 'for_profit'
-  | 'non_profit'
-  | 'government_entity'
-  | 'individual'
-  | '';
-
-
-export const BusinessTypeOptions: { label: string; value: BusinessType }[] = [
-  {
-    label: 'Choose business type',
-    value: '',
-  },
-  {
-    label: 'Individual',
-    value: 'individual',
-  },
-  {
-    label: 'For Profit',
-    value: 'for_profit',
-  },
-  {
-    label: 'Non Profit',
-    value: 'non_profit',
-  },
-  {
-    label: 'Government Entity',
-    value: 'government_entity',
-  },
-];
