+++ packages/webcomponents/src/components/business-forms/utils/business-form-types.ts
@@ -1,3 +1,5 @@
+import { BusinessType } from "../../../api/Business";
+
 export interface BusinessFormSubmitEvent {
   data: any;
   metadata?: BusinessFormEventMetaData;
@@ -59,3 +61,26 @@ export enum OwnerFormClickActions {
   addOwnerForm = 'addOwnerForm',
   updateOwner = 'updateOwner'
 }
+
+export const BusinessTypeOptions: { label: string; value: BusinessType | string }[] = [
+  {
+    label: 'Choose business type',
+    value: '',
+  },
+  {
+    label: 'Individual',
+    value: BusinessType.individual,
+  },
+  {
+    label: 'For Profit',
+    value: BusinessType.for_profit,
+  },
+  {
+    label: 'Non Profit',
+    value: BusinessType.non_profit,
+  },
+  {
+    label: 'Government Entity',
+    value: BusinessType.government_entity,
+  },
+];
